requirements.txt
streamlit
groq
python-dotenv
pandas
regex
